replace example-2 with empty template
