+++ contracts/empty-template/.gitignore
