+++ contracts/empty-template/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "example-2"
+name = "empty-template"
 version = "0.1.0"
 authors = ["Atomik Labs <https://atomiklabs.io>"]
 edition = "2018"
@@ -33,12 +33,12 @@ backtraces = ["cosmwasm-std/backtraces"]
 library = []
 
 [dependencies]
-cosmwasm-std = { version = "0.16.0" }
-cosmwasm-storage = { version = "0.16.0" }
-cw-storage-plus = "0.8.1"
-schemars = "0.8.3"
-serde = { version = "1.0.127", default-features = false, features = ["derive"] }
-thiserror = { version = "1.0.26" }
+cosmwasm-std = "0.16.2"
+cosmwasm-storage = "0.16.2"
+cw-storage-plus = "0.9.0"
+schemars = "0.8.4"
+serde = { version = "1.0.130", default-features = false, features = ["derive"] }
+thiserror = "1.0.29"
 
 [dev-dependencies]
-cosmwasm-schema = { version = "0.16.0" }
+cosmwasm-schema = { version = "0.16.2" }
