+++ contracts/empty-template/src/lib.rs
